requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
datasets==3.2.0
docker-pycreds==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
huggingface-hub==0.27.1
idna==3.10
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
pytube==15.0.0
pytubefix==8.1.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sentry-sdk==2.18.0
setproctitle==1.3.3
setuptools==75.5.0
six==1.16.0
smmap==5.0.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wandb==0.18.7
